Added Config.cpp to RHS and Vanila /Loadout
