+++ Addons/Manticore_Loadout/Inport.hpp
@@ -1,119 +1,6 @@
 class LandVehicle;
-///3CBVehicles
-//Light Attack
-class UK3CB_BAF_LandRover_WMIK_Milan_Green_A;
-class UK3CB_BAF_LandRover_WMIK_HMG_Green_A;
-class UK3CB_BAF_LandRover_WMIK_GPMG_Green_A;
-class UK3CB_BAF_LandRover_WMIK_GMG_Green_A;
-class UK3CB_BAF_LandRover_Soft_Green_A;
-class UK3CB_BAF_LandRover_Hard_Green_A;
-class UK3CB_BAF_LandRover_Snatch_Green_A;
-class UK3CB_BAF_LandRover_Amb_Green_A;
-class UK3CB_BAF_LandRover_Soft_MERT_A;
-class UK3CB_BAF_LandRover_Hard_MERT_A;
-class UK3CB_BAF_Coyote_Logistics_L134A1_W;
-class UK3CB_BAF_Coyote_Logistics_L111A1_W;
-class UK3CB_BAF_Coyote_Passenger_L134A1_W;
-class UK3CB_BAF_Coyote_Passenger_L111A1_W;
-class UK3CB_BAF_Jackal2_L111A1_W;
-class UK3CB_BAF_Jackal2_L134A1_W;
-//Car
-class UK3CB_BAF_Panther_GPMG_Green_A;
-//APC
-class UK3CB_BAF_FV432_Mk3_RWS_Green;
-class UK3CB_BAF_FV432_Mk3_GPMG_Green;
-//Heli
-class UK3CB_BAF_Wildcat_AH1_CAS_6A;
-class UK3CB_BAF_Wildcat_AH1_CAS_6B;
-class UK3CB_BAF_Wildcat_AH1_CAS_6C;
-class UK3CB_BAF_Wildcat_AH1_CAS_6D;
-class UK3CB_BAF_Wildcat_AH1_HEL_6A;
-class UK3CB_BAF_Wildcat_AH1_CAS_8A;
-class UK3CB_BAF_Wildcat_AH1_CAS_8B;
-class UK3CB_BAF_Wildcat_AH1_CAS_8C;
-class UK3CB_BAF_Wildcat_AH1_CAS_8D;
-class UK3CB_BAF_Wildcat_AH1_HEL_8A;
-class UK3CB_BAF_Wildcat_AH1_TRN_8A;
-class UK3CB_BAF_Wildcat_HMA2_TRN_8A;
-class UK3CB_BAF_Merlin_HC4_18;
-class UK3CB_BAF_Merlin_HC4_18_GPMG;
-class UK3CB_BAF_Merlin_HC4_24;
-class UK3CB_BAF_Merlin_HC4_32;
-class UK3CB_BAF_Merlin_HC4_Cargo;
-class UK3CB_BAF_Merlin_HC4_CSAR;
-class UK3CB_BAF_Merlin_HM2_18;
-class UK3CB_BAF_Apache_AH1_DynamicLoadout;
-class UK3CB_BAF_Apache_AH1_DynamicLoadoutUnlimited;
-class UK3CB_BAF_Apache_AH1_AT;
-class UK3CB_BAF_Apache_AH1_CAS;
-class UK3CB_BAF_Apache_AH1;
-class UK3CB_BAF_Apache_AH1_JS;
-//Naval
-class UK3CB_BAF_RHIB_HMG;
-class UK3CB_BAF_RHIB_GPMG;
 
-///RHS
-//Car
-class rhsusf_m1043_w_s;
-class rhsusf_m1043_w_s_m2;
-class rhsusf_m1043_w_s_mk19;
-class rhsusf_m1045_w_s;
-class rhsusf_m998_w_s_2dr;
-class rhsusf_m998_w_s_2dr_halftop;
-class rhsusf_m998_w_s_2dr_fulltop;
-class rhsusf_m998_w_s_4dr;
-class rhsusf_m998_w_s_4dr_halftop;
-class rhsusf_m998_w_s_4dr_fulltop;
-//MRAP
-class rhsusf_CGRCAT1A2_usmc_wd;
-class rhsusf_CGRCAT1A2_M2_usmc_wd;
-class rhsusf_CGRCAT1A2_Mk19_usmc_wd;
-class rhsusf_M1220_usarmy_wd;
-class rhsusf_M1220_M2_usarmy_wd;
-class rhsusf_M1220_M153_M2_usarmy_wd;
-class rhsusf_M1220_MK19_usarmy_wd;
-class rhsusf_M1230_M2_usarmy_wd;
-class rhsusf_M1230_MK19_usarmy_wd;
-class rhsusf_M1230a1_usarmy_wd;
 
-///Vanila
-//Light Attack
-class B_LSV_01_unarmed_F;
-class B_LSV_01_armed_F;
-class B_LSV_01_AT_F;
-class B_CTRG_LSV_01_light_F;
-class B_Quadbike_01_F;
-//Heli
-class I_Heli_light_03_dynamicLoadout_F;
-class I_Heli_light_03_unarmed_F;
-class I_Heli_Transport_02_F;
-class B_Heli_Light_01_F;
-class B_Heli_Light_01_dynamicLoadout_F;
-class B_Heli_Light_01_stripped_F;
-//APC IFV
-class Manticore_Vehicle_Badger_IFV;
-class Manticore_Vehicle_Namer;
-class Manticore_Vehicle_Nemmera;
-//AA
-class Manticore_Vehicle_Bardelas;
-//Tank
-class Manticore_Vehicle_Leopard;
-class I_LT_01_cannon_F;
-class I_LT_01_AT_F;
-class I_LT_01_scout_F;
-class B_AFV_Wheeled_01_cannon_F;
-class B_AFV_Wheeled_01_up_cannon_F;
-//Artillery
-class B_MBT_01_arty_F;
-class B_MBT_01_mlrs_F;
-//Naval
-class Manticore_Vehicle_speedboat_Minigun;
-//Drone
-class B_UAV_01_F;
-class B_T_UAV_03_dynamicLoadout_F;
-class B_UAV_05_F;
-class B_UGV_01_F;
-class B_UGV_01_rcws_F;
 
 ///Crates
 //"A3_Weapons_F_Ammoboxes"//
