+++ Addons/Manticore_Loadout/RHS/RHSInport.hpp
@@ -1,0 +1,29 @@
+//Car
+class rhsusf_m1043_w_s;
+class rhsusf_m1043_w_s_m2;
+class rhsusf_m1043_w_s_mk19;
+class rhsusf_m1045_w_s;
+class rhsusf_m998_w_s_2dr;
+class rhsusf_m998_w_s_2dr_halftop;
+class rhsusf_m998_w_s_2dr_fulltop;
+class rhsusf_m998_w_s_4dr;
+class rhsusf_m998_w_s_4dr_halftop;
+class rhsusf_m998_w_s_4dr_fulltop;
+//MRAP
+class rhsusf_CGRCAT1A2_usmc_wd;
+class rhsusf_CGRCAT1A2_M2_usmc_wd;
+class rhsusf_CGRCAT1A2_Mk19_usmc_wd;
+class rhsusf_M1220_usarmy_wd;
+class rhsusf_M1220_M2_usarmy_wd;
+class rhsusf_M1220_M153_M2_usarmy_wd;
+class rhsusf_M1220_MK19_usarmy_wd;
+class rhsusf_M1230_M2_usarmy_wd;
+class rhsusf_M1230_MK19_usarmy_wd;
+class rhsusf_M1230a1_usarmy_wd;
+//APC
+class rhsusf_m113_usarmy_unarmed;
+class rhsusf_m113_usarmy;
+class rhsusf_m113_usarmy_M240;
+class rhsusf_m113_usarmy_MK19;
+class rhsusf_m113_usarmy_supply;
+class rhsusf_m113_usarmy_medical;
