+++ Addons/Manticore_Loadout/Vanila/VanilaInport.hpp
@@ -1,0 +1,37 @@
+//Light Attack
+class B_LSV_01_unarmed_F;
+class B_LSV_01_armed_F;
+class B_LSV_01_AT_F;
+class B_CTRG_LSV_01_light_F;
+class B_Quadbike_01_F;
+//Heli
+class I_Heli_light_03_dynamicLoadout_F;
+class I_Heli_light_03_unarmed_F;
+class I_Heli_Transport_02_F;
+class B_Heli_Light_01_F;
+class B_Heli_Light_01_dynamicLoadout_F;
+class B_Heli_Light_01_stripped_F;
+//APC IFV
+class Manticore_Vehicle_Badger_IFV;
+class Manticore_Vehicle_Namer;
+class Manticore_Vehicle_Nemmera;
+//AA
+class Manticore_Vehicle_Bardelas;
+//Tank
+class Manticore_Vehicle_Leopard;
+class I_LT_01_cannon_F;
+class I_LT_01_AT_F;
+class I_LT_01_scout_F;
+class B_AFV_Wheeled_01_cannon_F;
+class B_AFV_Wheeled_01_up_cannon_F;
+//Artillery
+class B_MBT_01_arty_F;
+class B_MBT_01_mlrs_F;
+//Naval
+class Manticore_Vehicle_speedboat_Minigun;
+//Drone
+class B_UAV_01_F;
+class B_T_UAV_03_dynamicLoadout_F;
+class B_UAV_05_F;
+class B_UGV_01_F;
+class B_UGV_01_rcws_F;
